Refactor appWithTranslation to work with custom routes
We cannot use window.location.pathname in lngPathCorrector if the
route is a custom route, since window.location.pathname !== router.
pathname (instead, it is equal to router.asPath).  Since this is the
case, we cannot use shallow routing and a new page is requested,
resulting in a 404 error whenever the i18n.changeLanguage function is
invoked.

This commit changes appWithTranslation to instead use the withRouter
Next HoC, which exposes a router prop.  The router prop has pathname,
query, and asPath attributes which we can use instead of using
window.location.pathname.  We then modify query and asPath in
lngPathCorrector (as pathname will remain unchanged if we are only
changing the language) and shallow route if the asPath has changed
after correction (with query being modified, as necessary).